+++ src/test/kotlin/blackjack/service/HandsCalculatorTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.service
 
 import blackjack.domain.PokerCard
 import blackjack.domain.PokerSymbol
